Apply a clean-registry fixture for all the tests
